src/main/java/ru/job4j/todo/servlet/DoneServlet.java
package ru.job4j.todo.servlet;

import ru.job4j.todo.model.Task;
import ru.job4j.todo.service.HibernateService;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

public class DoneServlet extends HttpServlet {
    @Override
    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        int id = Integer.parseInt(req.getParameter("id"));
        HibernateService.instOf().setDone(new Task(id));
    }
}
